corrected the comments
