+++ srcs/execution/builtin/builtin_fcts/cd.c
@@ -6,17 +6,16 @@
 /*   By: mat <mat@student.42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/19 00:36:27 by mat               #+#    #+#             */
-/*   Updated: 2023/05/01 11:06:48 by mat              ###   ########.fr       */
+/*   Updated: 2023/05/01 11:22:35 by mat              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-static int	handle_chdir_err(char *const cd_arg)
+static void	handle_chdir_err(char *const cd_arg)
 {
 	print_error("%s: %s: %s: ", MINISHELL, CD_BUILTIN, cd_arg);
 	perror(EMPTY_STR);
-	return (EXIT_FAILURE);
 }
 
 static int	execute_cd(t_command *cmd_data)
@@ -26,20 +25,19 @@ static int	execute_cd(t_command *cmd_data)
 			cmd_data->command[1], &is_print);
 	int				ret_val;
 
-	ret_val = 0;
+	ret_val = EXIT_FAILURE;
 	check_pos(CHDIR);
-	if (cd_arg == NULL)
-		ret_val = 1;
-	else
+	if (cd_arg != NULL)
 	{
 		if (chdir(cd_arg) != -1)
 		{
 			update_cwd_var(cd_arg);
 			if (is_print == true)
 				print_pos();
+			ret_val = EXIT_SUCCESS;
 		}
 		else
-			ret_val = handle_chdir_err(cd_arg);
+			handle_chdir_err(cd_arg);
 	}
 	free(cd_arg);
 	is_print = false;
