+++ srcs/execution/builtin/exec_builtin.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   exec_builtin.c                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: mdorr <mdorr@student.42.fr>                +#+  +:+       +#+        */
+/*   By: mat <mat@student.42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/13 16:38:28 by rbroque           #+#    #+#             */
-/*   Updated: 2023/04/28 11:07:36 by mdorr            ###   ########.fr       */
+/*   Updated: 2023/05/01 11:26:19 by mat              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,7 +28,7 @@ static int	call_builtin(const t_builtin_mapper *map, t_command *cmd_data)
 	}
 	if (map[i].fct != NULL)
 		return (map[i].fct(cmd_data));
-	return (1);
+	return (EXIT_FAILURE);
 }
 
 void	exec_builtin(t_command *command)
