+++ srcs/execution/execution.c
@@ -6,57 +6,11 @@
 /*   By: mat <mat@student.42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/12 15:52:01 by rbroque           #+#    #+#             */
-/*   Updated: 2023/04/30 16:41:49 by mat              ###   ########.fr       */
+/*   Updated: 2023/05/01 11:29:56 by mat              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
-
-extern t_global	g_global;
-
-bool	is_folder(const char *path)
-{
-	struct stat	file_stat;
-
-	if (stat(path, &file_stat) == -1)
-	{
-		perror(STAT_ERROR);
-		free_manager();
-		exit(EXIT_FAILURE);
-	}
-	return (S_ISDIR(file_stat.st_mode));
-}
-
-static void	child_job(t_command *cmd_data, char *path)
-{
-	if (is_folder(path) == true)
-	{
-		g_global.last_ret_val = NO_ACCESS;
-		print_error("%s: %s: %s\n", MINISHELL, path, IS_DIR);
-		free_manager();
-		exit(g_global.last_ret_val);
-	}
-	else if (execve(path, cmd_data->command, cmd_data->env) == -1)
-		exit(g_global.last_ret_val);
-}
-
-static void	exec_binary(t_command *cmd_data, char *path)
-{
-	int	pid;
-	int	status;
-
-	if (path != NULL)
-	{
-		pid = fork();
-		if (pid == 0)
-			child_job(cmd_data, path);
-		else if (pid > 0)
-		{
-			wait(&status);
-			g_global.last_ret_val = extract_return_status(status);
-		}
-	}
-}
 
 static char	*get_path(t_command *cmd_data)
 {
